Refactor to app & pkg directories
