+++ pkg/flexadapter/examples/simplenfs-flexdriver/deploy/kubernetes/csi-attacher-simplenfs-flexdriver.yaml
