+++ pkg/flexadapter/examples/simplenfs-flexdriver/deploy/kubernetes/csi-nodeplugin-simplenfs-flexdriver.yaml
