+++ pkg/flexadapter/examples/simplenfs-flexdriver/examples/kubernetes/nginx.yaml
