+++ pkg/flexadapter/identityserver.go
@@ -17,7 +17,7 @@ limitations under the License.
 package flexadapter
 
 import (
-	"github.com/kubernetes-csi/drivers/csi-common"
+	"github.com/kubernetes-csi/drivers/pkg/csi-common"
 )
 
 type identityServer struct {
